+++ apps/learn-card-app/src/components/new-ai-session/LearnCardAiChatBot/FinishSessionButton.tsx
@@ -1,7 +1,11 @@
+import { useHistory } from 'react-router-dom';
 import { useStore } from '@nanostores/react';
-import { useGetCredentialList, useSyncConsentFlow } from 'learn-card-base';
+import { useGetCredentialList, useModal, useSyncConsentFlow } from 'learn-card-base';
 
+import X from '../../svgs/X';
 import TickSquareIcon from 'learn-card-base/svgs/TickSquareIcon';
+import { AiInsightsIconMonochrome } from 'learn-card-base/svgs/wallet/AiInsightsIcon';
+import { AiSessionsIconMonochrome } from 'learn-card-base/svgs/wallet/AiSessionsIcon';
 
 import {
     currentThreadId,
@@ -9,13 +13,32 @@ import {
     sessionEnded,
     finishSession,
     isEndingSession,
+    closeInsightsSession,
+    resetChatStores,
+    isTyping,
+    isLoading,
+    messages,
+    planReadyThread,
 } from 'learn-card-base/stores/nanoStores/chatStore';
+import { chatBotStore } from '../../../stores/chatBotStore';
+import { AiSessionMode } from '../newAiSession.helpers';
+import { ChatBotQuestionsEnum } from '../NewAiSessionChatBot/newAiSessionChatbot.helpers';
 
 const FinishSessionButton: React.FC = () => {
+    const history = useHistory();
+    const { closeAllModals } = useModal();
     const $sessionEnded = useStore(sessionEnded);
     const $currentThreadId = useStore(currentThreadId);
     const $threads = useStore(threads);
     const $isEndingSession = useStore(isEndingSession);
+    const $isTyping = useStore(isTyping);
+    const $isLoading = useStore(isLoading);
+    const $planReadyThread = useStore(planReadyThread);
+    const $messages = useStore(messages);
+
+    const mode = chatBotStore.useTracked.mode();
+    const qa = chatBotStore.useTracked.chatBotQA();
+    const promptTitle = qa.find(q => q.type === ChatBotQuestionsEnum.TopicSelection)?.answer;
 
     const { refetch: fetchNewContractCredentials } = useSyncConsentFlow();
     const { refetch: fetchTopics } = useGetCredentialList('AI Topic');
@@ -23,16 +46,75 @@ const FinishSessionButton: React.FC = () => {
     // Check if session has ended - either via atom or by checking for summary credentials
     const thread = $threads.find(t => t.id === $currentThreadId);
     const hasSessionEnded = $sessionEnded || (thread?.summaries && thread.summaries.length > 0);
+    const isSessionTyping = $isTyping;
 
     if (hasSessionEnded || $isEndingSession) return <></>;
 
-    const handleFinish = (e: React.MouseEvent<HTMLButtonElement>) => {
+    const handleFinish = async (e: React.MouseEvent<HTMLButtonElement>) => {
         e.preventDefault();
+
+        if (mode === AiSessionMode.insights) {
+            closeAllModals();
+            await closeInsightsSession();
+            chatBotStore.set.resetStore();
+            resetChatStores();
+            history.push('/ai/insights');
+            return;
+        }
+
+        // if the plan is ready and we are no longer loading, reset the chat stores
+        if ($planReadyThread && !$isLoading && $messages.length === 0) {
+            resetChatStores();
+            chatBotStore.set.resetStore();
+            closeAllModals();
+            return;
+        }
+
         finishSession(async () => {
             await fetchNewContractCredentials();
             await fetchTopics();
         });
     };
+
+    if (mode === AiSessionMode.insights) {
+        return (
+            <div className="min-h-[75px] flex justify-between items-center gap-[15px] p-[15px] absolute top-[0px] left-[0px] w-full bg-white shadow-box-bottom z-[100000] safe-area-inset-top">
+                <div className="flex items-center justify-center gap-1">
+                    <AiInsightsIconMonochrome className="w-[24px] h-[24px] mb-1" />
+                    <p className="text-sm flex items-center justify-center font-[600] leading-[24px] tracking-[0.25px] text-grayscale-600">
+                        {promptTitle}
+                    </p>
+                </div>
+                {!isSessionTyping && (
+                    <button
+                        onClick={handleFinish}
+                        className="w-[24px] h-[24px] flex items-center justify-center text-grayscale-600"
+                    >
+                        <X />
+                    </button>
+                )}
+            </div>
+        );
+    }
+
+    return (
+        <div className="min-h-[75px] flex justify-between items-center gap-[15px] p-[15px] absolute top-[0px] left-[0px] w-full bg-white shadow-box-bottom z-[100000] safe-area-inset-top">
+            <div className="flex items-center justify-center gap-1 text-grayscale-900 font-semibold">
+                <AiSessionsIconMonochrome className="w-[24px] h-[24px] mr-2" />
+                AI Session
+                {/* <p className="text-sm flex items-center justify-center font-[600] leading-[24px] tracking-[0.25px] text-grayscale-600">
+                        {promptTitle}
+                    </p> */}
+            </div>
+
+            <button
+                onClick={handleFinish}
+                className="w-[24px] h-[24px] flex items-center justify-center text-grayscale-600"
+            >
+                <X />
+            </button>
+        </div>
+    );
 
     return (
         <div className="flex justify-center items-center p-[15px] absolute top-[0px] left-[0px] w-full bg-white shadow-box-bottom z-[100000] safe-area-inset-top">