+++ packages/learn-card-base/src/react-query/mutations/ai-passport.ts
@@ -1,14 +1,20 @@
 import { UnsignedVC } from '@learncard/types';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { useModal, useToast, ToastTypeEnum, UploadTypesEnum } from 'learn-card-base';
+import {
+    useModal,
+    useToast,
+    ToastTypeEnum,
+    UploadTypesEnum,
+    LEARNCARD_AI_URL,
+} from 'learn-card-base';
 
 export const usePreloadAssessment = () => {
     const queryClient = useQueryClient();
 
     return useMutation({
         mutationFn: async ({ did, summaryCredential }: { did: string; summaryCredential: any }) => {
-            const res = await fetch(`https://api.learncloud.ai/assessment?did=${did}`, {
+            const res = await fetch(`${LEARNCARD_AI_URL}/assessment?did=${did}`, {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
                 body: JSON.stringify({ summaryCredential }),
@@ -38,7 +44,7 @@ type FinishAssessmentPayload = {
 export const useFinishAssessmentMutation = () => {
     return useMutation({
         mutationFn: async ({ did, assessmentQA, session, sessionUri }: FinishAssessmentPayload) => {
-            const response = await fetch(`https://api.learncloud.ai/finish-assessment?did=${did}`, {
+            const response = await fetch(`${LEARNCARD_AI_URL}/finish-assessment?did=${did}`, {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
                 body: JSON.stringify({ assessmentQA, session, sessionUri }),
@@ -69,7 +75,7 @@ export const useUploadFileMutation = (fileType: UploadTypesEnum) => {
         }) => {
             try {
                 const response = await fetch(
-                    `https://api.learncloud.ai/credentials/parse-file?did=${did}`,
+                    `${LEARNCARD_AI_URL}/credentials/parse-file?did=${did}`,
                     {
                         method: 'POST',
                         headers: { 'Content-Type': 'application/json' },